# api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: NextJourney REST API
  description: "API REST de NextJourney para gestionar usuarios, destinos, viajes,\
    \ imágenes y reseñas."
  version: v1
servers:
- url: https://localhost:8443
  description: Generated server url
...
paths:
...
  /api/v1/auth/register:
    post:
      tags:
      - Autenticación
      operationId: register
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UserDTO"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/auth/refresh:
    post:
      tags:
      - Autenticación
      operationId: refresh
      parameters:
      - name: RefreshToken
        in: cookie
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /api/v1/auth/login:
    post:
      tags:
      - Autenticación
      operationId: login
      parameters:
      - name: user
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/UserDTO"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /api/v1/users/{id}:
    delete:
      tags:
      - Usuarios
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
    ImageDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        password:
          type: string
          writeOnly: true
        image:
          $ref: "#/components/schemas/ImageDTO"
        roles:
          type: array
          items:
            type: string
...
    AuthResponse:
      type: object
      properties:
        status:
          type: string
          enum:
          - SUCCESS
          - FAILURE
        message:
          type: string
        error:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      in: header
      scheme: bearer
      bearerFormat: JWT